removed select timeouts in socket_write_exact and socket_read_exect; removed WriteExact

+++ slots.h
@@ -30,7 +30,7 @@ typedef struct _repeaterslot
 {
 	SOCKET server;
 	SOCKET viewer;
-	unsigned long timestamp;
+	uint32_t timestamp;
 	uint32_t code;
 	unsigned char challenge[CHALLENGESIZE];
